Create README.md
Auxiliary script to restore a backup created by raspiBackup in a simple dialog-guided way.

It can simply be started without options.
Then a query appears if the last backup should be restored. (y/N)
If (N) an appropriate backup can be selected from a list. 
Then select the target medium and raspiBackup does the rest.

Possible Options --last or --select